% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fit1}
\alias{fit1}
\title{Random forest fitted object from Caret on Sonar data}
\format{A Caret train object}
\usage{
fit1
}
\description{
Caret was run using 10 fold repeated cross validation on the Sonar data
with random forest used to predict the response variable.
}
\keyword{datasets}
