% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fit2}
\alias{fit2}
\title{Gradient boosted machines fitted object from Caret on Sonar data}
\format{A Caret train object}
\usage{
fit2
}
\description{
Caret was run using 10 fold repeated cross validation on Sonar data
with GBM used to predict the response variable.
}
\keyword{datasets}
