% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fit3}
\alias{fit3}
\title{Random forest fitted object from Caret on Sonar data with log-likelihood objective function}
\format{A Caret train object}
\usage{
fit3
}
\description{
Caret was run using 10 fold repeated cross validation on the Sonar data
using random forest to predict the response variable. Log-likelihood was set
to be the objective function to select the best model from cross validation.
}
\keyword{datasets}
