% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{im_fit}
\alias{im_fit}
\title{Random forest fitted object from Caret on simulated imbalanced data}
\format{A Caret train object}
\usage{
im_fit
}
\description{
Caret was run using 10 fold repeated cross validation on the Sonar data
with random forest to predict the response variable.
}
\keyword{datasets}
