% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{preds}
\alias{preds}
\title{Predictions from gbm on the Sonar test data}
\format{A data frame with 51 rows as points and 4 variables}
\usage{
preds
}
\description{
The Sonar data was split into training (157 points) and 
testing (51 points), a gbm model was fitted using Caret 
on the training data. Then these are the predicted 
probabilities of the model on the test data.
}
\keyword{datasets}
