% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predsc}
\alias{predsc}
\title{Predictions from gbm and random forest on the Sonar test data}
\format{A data frame with 102 rows as points and 4 variables}
\usage{
predsc
}
\description{
The Sonar data was split into training (157 points) and 
testing (51 points), a gbm model was fitted using Caret 
on the training data. Then these are the predicted 
probabilities of the model on the test data. A random
forest model was then fit and tested in the same manner. 
The probabilities and ground truth labels were combined 
horizontally for further analysis.
}
\keyword{datasets}
