% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sonar}
\alias{Sonar}
\title{Sonar data}
\format{A data frame with 208 rows as points and 61 variables}
\usage{
Sonar
}
\description{
The Sonar data consist of 208 data points collected on 60 predictors. 
The goal is to predict the two classes M for metal cylinder or R for rock.
This data has been obtained from the 'mlbench' package. Response variable
is in the Class column.
}
\keyword{datasets}
